+++ core/profile_api.php
@@ -80,7 +80,7 @@ function profile_create( $p_user_id, $p_platform, $p_os, $p_os_build, $p_descrip
 	$query = "INSERT INTO {user_profile}
 				    ( user_id, platform, os, os_build, description )
 				  VALUES
-				    ( " . db_param() . ', ' . db_param() . ', ' . db_param() . ', ' . db_param() . ', ' . db_param() . ' )';
+				    ( %d, %s, %s, %s, %s )";
 	db_query_bound( $query, array( $p_user_id, $p_platform, $p_os, $p_os_build, $p_description ) );
 
 	return db_insert_id( $t_user_profile_table );
@@ -105,8 +105,7 @@ function profile_delete( $p_user_id, $p_profile_id ) {
 	}
 
 	# Delete the profile
-	$query = "DELETE FROM {user_profile}
-				  WHERE id=" . db_param() . " AND user_id=" . db_param();
+	$query = 'DELETE FROM {user_profile} WHERE id=%d AND user_id=%d';
 	db_query_bound( $query, array( $c_profile_id, $c_user_id ) );
 
 	# db_query errors on failure so:
@@ -151,11 +150,11 @@ function profile_update( $p_user_id, $p_profile_id, $p_platform, $p_os, $p_os_bu
 
 	# Add item
 	$query = "UPDATE {user_profile}
-				  SET platform=" . db_param() . ",
-				  	  os=" . db_param() . ",
-					  os_build=" . db_param() . ",
-					  description=" . db_param() . "
-				  WHERE id=" . db_param() . " AND user_id=" . db_param();
+				  SET platform=%s,
+				  	  os=%s,
+					  os_build=%s,
+					  description=%s
+				  WHERE id=%d AND user_id=%d";
 	$result = db_query_bound( $query, array( $p_platform, $p_os, $p_os_build, $p_description, $c_profile_id, $c_user_id ) );
 
 	# db_query errors on failure so:
@@ -172,8 +171,7 @@ function profile_get_row( $p_user_id, $p_profile_id ) {
 	$c_user_id = db_prepare_int( $p_user_id );
 	$c_profile_id = db_prepare_int( $p_profile_id );
 
-	$query = "SELECT * FROM {user_profile}
-				  WHERE id=" . db_param() . " AND user_id=" . db_param();
+	$query = 'SELECT * FROM {user_profile} WHERE id=%d AND user_id=%d';
 	$result = db_query_bound( $query, array( $c_profile_id, $c_user_id ) );
 
 	return db_fetch_array( $result );
@@ -188,7 +186,7 @@ function profile_get_row( $p_user_id, $p_profile_id ) {
 function profile_get_row_direct( $p_profile_id ) {
 	$c_profile_id = db_prepare_int( $p_profile_id );
 
-	$query = "SELECT * FROM {user_profile} WHERE id=" . db_param();
+	$query = "SELECT * FROM {user_profile} WHERE id=%d";
 	$result = db_query_bound( $query, array( $c_profile_id ) );
 
 	return db_fetch_array( $result );
@@ -202,9 +200,7 @@ function profile_get_row_direct( $p_profile_id ) {
 function profile_get_all_rows( $p_user_id ) {
 	$c_user_id = db_prepare_int( $p_user_id );
 
-	$query = "SELECT * FROM {user_profile}
-				  WHERE user_id=" . db_param() . "
-				  ORDER BY platform, os, os_build";
+	$query = "SELECT * FROM {user_profile} WHERE user_id=%d ORDER BY platform, os, os_build";
 	$t_result = db_query_bound( $query, array( $c_user_id ) );
 
 	$t_rows = array();
@@ -256,7 +252,7 @@ function profile_get_field_all_for_user( $p_field, $p_user_id = null ) {
 	}
 
 	$query = "SELECT DISTINCT $c_field FROM {user_profile}
-				  WHERE ( user_id=" . db_param() . " ) OR ( user_id = 0 )
+				  WHERE ( user_id=%d ) OR ( user_id = 0 )
 				  ORDER BY $c_field";
 	$result = db_query_bound( $query, array( $c_user_id ) );
 
@@ -302,7 +298,7 @@ function profile_get_all_for_project( $p_project_id ) {
 function profile_get_default( $p_user_id ) {
 	$c_user_id = db_prepare_int( $p_user_id );
 
-	$query = "SELECT default_profile FROM {user_pref} WHERE user_id=" . db_param();
+	$query = 'SELECT default_profile FROM {user_pref} WHERE user_id=%d';
 	$result = db_query_bound( $query, array( $c_user_id ) );
 
 	$t_default_profile = db_result( $result );
